[GH-#104] Changed the mechanism of the chain state persistence
