+++ apps/aecore/apps/aecore/logs/2017-10-19_info.log
@@ -1,0 +1,102 @@
+10:39:27.463 [info] Mined block #1 with a difficulty target of 1
+10:39:27.489 [info] Added block #1 with a hash of
+907CCDF87588819EF410E345E8E7C8100DE879B155F9713653876A33FE2CA419 to the chain,
+total tokens in the chain - 100
+10:39:32.761 [info] Generated in 5267744 microsecs
+10:39:32.762 [info] Verified in 430 microsecs
+10:39:41.192 [info] Mined block #2 with a difficulty target of 1
+10:39:41.193 [info] Added block #2 with a hash of
+2B214039FB4963B52E4C329C624EA836AECC7A004E88D9183E80BCA00812CB33 to the chain,
+total tokens in the chain - 200
+10:39:42.022 [info] Mined block #3 with a difficulty target of 1
+10:39:42.022 [info] Added block #3 with a hash of
+8D983EC4D3A6F6B5384D478E973435005B2BEDEACFA501F3F2CD8724648BE131 to the chain,
+total tokens in the chain - 300
+10:39:43.435 [info] Added block #4 with a hash of
+1018E2A008B5973CB8E8472C310B821A9B709361924A03A287E6F8EA79976FB4 to the chain,
+total tokens in the chain - 300
+10:41:15.296 [info] Mined block #1 with a difficulty target of 1
+10:41:15.314 [info] Added block #1 with a hash of
+9E33E4B5AA9AC707F5BD2882052C565C5AC945B95A984028FA38479FECD33E4E to the chain,
+total tokens in the chain - 100
+10:41:20.071 [info] Generated in 4748612 microsecs
+10:41:20.071 [info] Verified in 323 microsecs
+10:41:23.394 [info] Mined block #2 with a difficulty target of 1
+10:41:23.395 [info] Added block #2 with a hash of
+9F1E8FCD498ED9545B73B3EE4DCEAE5D50A81439320C3DBE7B3BE4174E5D37F8 to the chain,
+total tokens in the chain - 200
+10:41:23.515 [info] Mined block #3 with a difficulty target of 1
+10:41:23.516 [info] Added block #3 with a hash of
+692C33CB1855C580813EA13B3857A7421031E054EEDA90BC30F880953C281645 to the chain,
+total tokens in the chain - 300
+10:41:23.975 [info] Added block #4 with a hash of
+36B716639D1DBB63D24D7821A5BE2F21A85DF1167070B829E00369F01B7E4919 to the chain,
+total tokens in the chain - 300
+10:41:56.813 [info] Mined block #1 with a difficulty target of 1
+10:41:56.836 [info] Added block #1 with a hash of
+121627769B53C1A24C2759C08CF6F51564CDC46126ACF0AB5D0EBAD8A49E5934 to the chain,
+total tokens in the chain - 100
+10:42:02.177 [info] Generated in 5340364 microsecs
+10:42:02.178 [info] Verified in 365 microsecs
+10:42:06.594 [info] Mined block #2 with a difficulty target of 1
+10:42:06.595 [info] Added block #2 with a hash of
+FA34E6099A21319277CE26385F0BAD1BE28FE06FD3291ACC7B93BC1B9C9AEE7A to the chain,
+total tokens in the chain - 200
+10:42:10.601 [info] Mined block #3 with a difficulty target of 1
+10:42:10.603 [info] Added block #3 with a hash of
+4165851B078082CCA7837BF168F351CEC40D8203521C990B6F3BEA4266C7AE65 to the chain,
+total tokens in the chain - 300
+10:42:15.976 [info] Added block #4 with a hash of
+7BF88E57909A56D6CF54D8974711273B0A99F19747DB0C36CA2175FE1C28E636 to the chain,
+total tokens in the chain - 300
+10:51:38.830 [info] Mined block #1 with a difficulty target of 1
+10:51:38.847 [info] Added block #1 with a hash of
+AD2D023A5C0E8F321DBA5BB0EA38C4AA28447F182F9EF6DF4BD12531BCD43503 to the chain,
+total tokens in the chain - 100
+10:51:43.793 [info] Generated in 4939239 microsecs
+10:51:43.794 [info] Verified in 705 microsecs
+10:51:45.034 [info] Mined block #2 with a difficulty target of 1
+10:51:45.035 [info] Added block #2 with a hash of
+DDC061D01230BC599E418C41F139C60C7E3D048877F80DC28EF4D0F2A5192947 to the chain,
+total tokens in the chain - 200
+10:51:47.916 [info] Mined block #3 with a difficulty target of 1
+10:51:47.917 [info] Added block #3 with a hash of
+D3339590552E02F116BB071FF8EBA116CE7820CD69D48C241A650A78018AD080 to the chain,
+total tokens in the chain - 300
+10:51:52.281 [info] Added block #4 with a hash of
+A80F4050C474FBEA63F9031287475C7DDFB77F334C8C7545CDED23925846D90B to the chain,
+total tokens in the chain - 300
+10:53:25.286 [info] Mined block #1 with a difficulty target of 1
+10:53:25.310 [info] Added block #1 with a hash of
+5690040961D142F4265F2569CFC1C764CF655D2F19FF955382E42B6F4C64F2BF to the chain,
+total tokens in the chain - 100
+10:53:30.194 [info] Generated in 4879240 microsecs
+10:53:30.195 [info] Verified in 456 microsecs
+10:53:30.333 [info] Mined block #2 with a difficulty target of 1
+10:53:30.334 [info] Added block #2 with a hash of
+6A4ACCA2DF58D7B7041F3632A7F4F588B15C7619F3E8EE67280390279ADB8F7F to the chain,
+total tokens in the chain - 200
+10:53:31.577 [info] Mined block #3 with a difficulty target of 1
+10:53:31.579 [info] Added block #3 with a hash of
+0DC178A0C9787D92CAEEC65A35ACBA989A74041BB1F506FE81F4CE025E46E4F5 to the chain,
+total tokens in the chain - 300
+10:53:36.817 [info] Added block #4 with a hash of
+2CDE8DADDAD52DE2C45BF4790BF599AE960E4E70B5E665753047EA491A3FEE8B to the chain,
+total tokens in the chain - 300
+11:05:57.111 [info] Mined block #1 with a difficulty target of 1
+11:05:57.129 [info] Added block #1 with a hash of
+558CBC4F5421990C6EED7C4FFB65AB8FFD150422B202FF01139B25FE9C242179 to the chain,
+total tokens in the chain - 100
+11:06:02.413 [info] Generated in 5278613 microsecs
+11:06:02.413 [info] Verified in 399 microsecs
+11:06:06.966 [info] Mined block #2 with a difficulty target of 1
+11:06:06.967 [info] Added block #2 with a hash of
+1E1CC4EC196BCD3026428B5E7F9E0271A3E0DE9E5D850C3B7DC6AFCE0193E738 to the chain,
+total tokens in the chain - 200
+11:06:18.151 [info] Mined block #3 with a difficulty target of 1
+11:06:18.152 [info] Added block #3 with a hash of
+A0E166BA1B93D9D1491EE9417C99A1D6998AD95A78A76E7ADE9C2FD00D9EA116 to the chain,
+total tokens in the chain - 300
+11:06:22.077 [info] Added block #4 with a hash of
+25C845E291F120072F09881541CE098716F47B0694A174E18C7F72A7AA28EA11 to the chain,
+total tokens in the chain - 300
