+++ apps/aecore/lib/aecore.ex
@@ -7,8 +7,7 @@ defmodule Aecore do
       Aecore.Chain.Worker.Supervisor,
       Aecore.Miner.Worker.Supervisor,
       Aecore.Txs.Pool.Worker.Supervisor,
-      Aecore.Peers.Worker.Supervisor,
-      Aecore.Persistence.Worker.Supervisor
+      Aecore.Peers.Worker.Supervisor
     ]
 
     Supervisor.start_link(children, strategy: :one_for_one)
